app/requirements.txt
langgraph
openai
qdrant-client
tavily-python
llama-index
fastapi
uvicorn
python-dotenv
httpx
loguru